main_dow30_randomTic_discrete_256neu_ppo_3m_steps
While the training addressed a clear tendency to converge, it became distracted from the improvement and turn away from getting better at the end of the training. I think the problem come from the number of trades that increase cost of trading when the agent encounter with a downward trend. Try to ad number of trades as an item of the state.