main_dow30_randomTic_discrete_standadized_shortSell_4hiden_ppo4_2m_steps(ppo_52)
The idea of applying just only reward at the episode end with the training start from main_dow30_randomTic_discrete_standadized_shortSell_4hiden_ppo2-2_6m_steps seem not to be good. I return to the previous one with hold_period is 5 steps. I am considering the idea of an action space that not just decide the trading actions but also set the threshold for filtering the fundamental metrics